backend/requirements.txt
flask
flask-cors
werkzeug
python-docx
pypdf2
reportlab
python-dotenv
google-generativeai